CORE-8621 Refactor DiskResourceColumnModel size provider and cells into separate classes
